Adding tests & docs
